+++ src/nvlib/model/exporter/nv_html_reporter.py
@@ -8,15 +8,15 @@ import os
 
 from nvlib.model.converter.export_target_factory import ExportTargetFactory
 from nvlib.model.file.doc_open import open_document
+from nvlib.model.html.html_chapter_structure_board import HtmlChapterStructureBoard
 from nvlib.model.html.html_characters import HtmlCharacters
 from nvlib.model.html.html_element_notes import HtmlElementNotes
 from nvlib.model.html.html_items import HtmlItems
 from nvlib.model.html.html_locations import HtmlLocations
-from nvlib.model.html.html_plot_cards import HtmlPlotCards
+from nvlib.model.html.html_plot_line_board import HtmlPlotLineBoard
 from nvlib.model.html.html_plot_list import HtmlPlotList
 from nvlib.model.html.html_project_notes import HtmlProjectNotes
-from nvlib.model.html.html_section_cards import HtmlSectionCards
-from nvlib.model.html.html_stage_cards import HtmlStageCards
+from nvlib.model.html.html_story_structure_board import HtmlStoryStructureBoard
 from nvlib.model.html.html_timetable import HtmlTimetable
 
 
@@ -33,10 +33,10 @@ class NvHtmlReporter:
         HtmlLocations,
         HtmlItems,
         HtmlPlotList,
-        HtmlPlotCards,
+        HtmlPlotLineBoard,
         HtmlProjectNotes,
-        HtmlSectionCards,
-        HtmlStageCards,
+        HtmlChapterStructureBoard,
+        HtmlStoryStructureBoard,
         HtmlTimetable,
         ]