+++ src/nvlib/model/html/html_plot_line_board.py
@@ -4,14 +4,14 @@ Copyright (c) Peter Triesberger
 For further information see https://github.com/peter88213/novelibre
 License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
-from nvlib.model.html.html_cards import HtmlCards
+from nvlib.model.html.html_board import HtmlBoard
 from nvlib.novx_globals import PLOT_CARD_SUFFIX
 from nvlib.novx_globals import PL_ROOT
 from nvlib.nv_locale import _
 
 
-class HtmlPlotCards(HtmlCards):
-    """html plot line board representation."""
+class HtmlPlotLineBoard(HtmlBoard):
+    """html Plot line board representation."""
     DESCRIPTION = _('HTML Plot line board')
     SUFFIX = PLOT_CARD_SUFFIX